% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{buildInstance}
\alias{buildInstance}
\title{Plot statistics for suggestion of tweaked instances (population)}
\usage{
buildInstance(X.train, true.y, ntree = 50, epsiron = 0.1,
  type = c("randomForest", "xgboost"), .tryModel = FALSE)
}
\arguments{
\item{X.train}{,}

\item{true.y}{,}

\item{ntree}{,}

\item{epsiron}{,}

\item{type}{what type of plot should be drawn. Possible types are
\itemize{
 \item{"randomForest"}{ set ensemble trees to \code{\link{randomForest}}.}
 \item{"xgboost"}{ set ensemble trees to \code{\link{xgboost}}. }
}
\code{"randomForest"} is defult.}

\item{.tryModel}{a logical. If \code{.tryModel = TRUE}, Only build learner.}
}
\value{
a list of
\itemize{
 \item{"origin"}{ text. }
 \item{"scaled"}{ text. }
 \item{"forest"}{ text. }
 \item{"rules"}{ text. }
 \item{"esatisfy"}{ text. }
}
Default is "absoluteSum".
}
\description{
Plot statistics for suggestion of tweaked instances (population)
}
