% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chop}
\alias{chop}
\title{Remove the First or Last n of an Object}
\usage{
chop(X, n = 1)
}
\arguments{
\item{X}{a data.frame or a matrix}

\item{n}{an integer. number of row to be removed from last. Same as \code{ head(X, NROW(X) - n)}.}
}
\description{
Remove the First or Last n of an Object
}
\examples{
chop(iris, 140)

}
