% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataSplit}
\alias{dataSplit}
\title{Split dataset into train set and test set (for hold-out)}
\usage{
dataSplit(dataset, test.ratio = 0.1)
}
\arguments{
\item{dataset}{a data.frame or a matrix}

\item{test.ratio}{a numelic. Ratio of test data, must be (0, 1).}
}
\description{
Split dataset into train set and test set (for hold-out)
}
\examples{
x <- dataSplit(1:10, test.ratio = 0.3)
x <- dataSplit(mtcars, test.ratio = 0.3)

}
