% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descale.tweakedFeature.R
\name{descale.tweakedFeature}
\alias{descale.tweakedFeature}
\title{Restore tweaked instances to the original scale}
\usage{
descale.tweakedFeature(tweaked.X, scaled.X)
}
\arguments{
\item{tweaked.X}{an object returned by tweak() to be restored.}

\item{scaled.X}{a scaled matrixa or a scaled data.frame. Usually train set or test set.}
}
\value{
a list of
\itemize{
 \item{"original"}{ a data.frame of the same instances before being scalsed. }
 \item{"suggest"}{ a data.frame of instances tweakd in original scale. }
 \item{"diff"}{ a tibble of \code{(suggest - original)} }
}
}
\description{
Restore tweaked instances to the original scale
}
\examples{
\dontrun{
X <- iris[, 1:(ncol(iris)-1)]
scaled.X <- scale(X)
true.y <- iris[, ncol(iris)]

rf.iris <- randomForest(scaled.X, true.y, ntree=30)
es.rf <- set.eSatisfactory(forest.rf, ntree = 30, epsiron = 0.3, resample = TRUE)
tweaked <- tweak(es.rf, newdata= scaled.X, label.from = "spam", label.to = "nonspam",
                 .dopar = TRUE)
dt <- descale.tweakedFeature(tweaked, scaled.X)
}

}
