% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRules.R
\name{getRules}
\alias{getRules}
\title{parse a decision tree in randomForest into list of path as data.frame}
\usage{
getRules(forest, ktree = NULL, resample = FALSE)
}
\arguments{
\item{forest}{an object of ensemble trees to be parsed. See \code{\link{randomForest}} or \code{\link{xgboost}}.}

\item{ktree}{an integer. number of decision tree to be parsed. If ktree=NULL (default), all tree will be parsed.}

\item{resample}{Logical. If TRUE, trees are ramdomly selected. If FALSE, trees are selected according to head(ktree) from forest.}
}
\value{
a list of trees (list).
}
\description{
parse a decision tree in randomForest into list of path as data.frame
}
\examples{
\dontrun{
X <- iris[, 1:(ncol(iris)-1)]
true.y <- iris[, ncol(iris)]

rf.iris <- randomForest(X, true.y, ntree=30)
getRules(rf.iris, ktree=15)
}

}
