% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{learnModel}
\alias{learnModel}
\title{TBD}
\usage{
learnModel(X.train, true.y, ntree = 50, epsiron = 0.1,
  type = c("randomForest", "xgboost"), .tryModel = FALSE)
}
\arguments{
\item{X.train}{.}

\item{true.y}{.}

\item{ntree}{.}

\item{epsiron}{.}

\item{type}{.}

\item{.tryModel}{.}
}
\description{
TBD
}
