% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{plot.featureTweak}
\alias{plot.featureTweak}
\title{TBD}
\usage{
\method{plot}{featureTweak}(x, ..., k = NULL, type = "frequency")
}
\arguments{
\item{x}{.}

\item{...}{.}

\item{k}{.}

\item{type}{.}
}
\description{
TBD
}
