% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotSuggest}
\alias{plotSuggest}
\title{Plot individual suggestion (direction to modify)}
\usage{
plotSuggest(x, k = 1, .ordered = FALSE, .nonzero.only = FALSE,
  .invisible = FALSE)
}
\arguments{
\item{x}{an object returned by tweak().}

\item{k}{an integer. The tweaked direction of k-th instance will be plotted.}

\item{.ordered}{a logical. If \code{.ordered = TRUE}, features are sorted by suggestion.}

\item{.nonzero.only}{a logical. If \code{.nonzero.only = TRUE}, feature with suggestion = 0 is not be shown.}

\item{.invisible}{a logical. If \code{.invisible = TRUE}, plot won't be shown.}
}
\value{
a ggplot object (for replot).
}
\description{
Plot individual suggestion (direction to modify)
}
\examples{
\dontrun{
X <- iris[, 1:(ncol(iris)-1)]
true.y <- iris[, ncol(iris)]

rf.iris <- randomForest(X, true.y, ntree=30)
es.rf <- set.eSatisfactory(forest.rf, ntree = 30, epsiron = 0.3, resample = TRUE)
tweaked <- tweak(es.rf, newdata= X.test, label.from = "spam", label.to = "nonspam",
                 .dopar = TRUE)

plotSuggest(tweaked, k = 1)
plotSuggest(tweaked, k = 1, .ordered = TRUE, .nonzero.only = TRUE)
}

}
