% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{predict.modelTweakFeature}
\alias{predict.modelTweakFeature}
\title{TBD}
\usage{
\method{predict}{modelTweakFeature}(object, ..., newdata, label.from, label.to)
}
\arguments{
\item{object}{.}

\item{...}{.}

\item{newdata}{.}

\item{label.from}{.}

\item{label.to}{.}
}
\description{
TBD
}
