% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescale}
\alias{rescale}
\title{Scale a matrix with parameters of a scaled matix}
\usage{
rescale(X, scaled)
}
\arguments{
\item{X}{a matrix or a data.frame to be scaled}

\item{scaled}{a scaled matrixa or a scaled data.frame}
}
\value{
a tibble
}
\description{
Scale a matrix with parameters of a scaled matix
}
\details{
X is scaled by \code{center = attr(scaled, "scaled:center")} and \code{scale = attr(scaled, "scaled:scale")}
}
\examples{
iris.scaled <- scale(iris[1:100,-5])


}
