% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{tweak}
\alias{tweak}
\title{calcurate suggention for each inputs based on prediction}
\usage{
tweak(obj, newdata, label.from, label.to, .dopar = TRUE)
}
\arguments{
\item{forest}{a randomForest object.}

\item{ntree}{an integer. Number of decision tree to be parsed. If ntree=NULL (default), all tree will be parsed.}

\item{resample}{Logical. If reample=TRUE, trees are ramdomly selected. If FALSE, trees are selected according to head(ntree) from forest.}

\item{epsiron}{a numeric. Amount of "tolerance". All standardized feature value changes from threashold with this tolerance.}
}
\value{
a list of trees (list).
}
\description{
calcurate suggention for each inputs based on prediction
}
\examples{
\dontrun{
X <- iris[, 1:(ncol(iris)-1)]
true.y <- iris[, ncol(iris)]

rf.iris <- randomForest(X, true.y, ntree=30)
es.rf <- set.eSatisfactory(forest.rf, ntree = 30, epsiron = 0.3, resample = TRUE)
}

}
