% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRules.R
\name{getRules}
\alias{getRules}
\title{parse a decision tree in randomForest into list of path as data.frame}
\usage{
getRules(forest, ntree = NULL, resample = FALSE)
}
\arguments{
\item{forest}{a randomForest object to be parsed}

\item{ntree}{an integer. number of decision tree to be parsed. If ntree=NULL (default), all tree will be parsed.}

\item{resample}{Logical. If TRUE, trees are ramdomly selected. If FALSE, trees are selected according to head(ntree) from forest.}
}
\value{
a list of trees (list).
}
\description{
parse a decision tree in randomForest into list of path as data.frame
}
\examples{
\dontrun{
X <- iris[, 1:(ncol(iris)-1)]
true.y <- iris[, ncol(iris)]

rf.iris <- randomForest(X, true.y, ntree=30)
getRules(rf.iris, ntree=15)
}

}
