% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.tweaked.suggestion}
\alias{plot.tweaked.suggestion}
\title{Plot statistics for suggestion of tweaked instances (population)}
\usage{
\method{plot}{tweaked.suggestion}(x, ..., type = c("absoluteSum", "direction",
  "frequency"), .invisible = FALSE)
}
\arguments{
\item{x}{an object returned by tweak().}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{type}{what type of plot should be drawn. Possible types are
\itemize{
 \item{"absoluteSum"}{ draw baplot for absolute sum of suggestions. }
 \item{"direction"}{ draw boxplot for suggestions among instances. }
 \item{"frequency"}{ draw baplot for total number of tweaked among instances. }
}
\code{"absoluteSum"} is defult.}

\item{.invisible}{a logical. If \code{.invisible = TRUE}, plot won't be shown.}
}
\value{
a list of
\itemize{
 \item{"stats"}{ a data.frame of statistics for each feature. }
 \item{"plot"}{ a ggplot object (for replot). }
}
Default is "absoluteSum".
}
\description{
Plot statistics for suggestion of tweaked instances (population)
}
\examples{
\dontrun{
X <- iris[, 1:(ncol(iris)-1)]
scaled.X <- scale(X)
true.y <- iris[, ncol(iris)]

rf.iris <- randomForest(scaled.X, true.y, ntree=30)
es.rf <- set.eSatisfactory(forest.rf, ntree = 30, epsiron = 0.3, resample = TRUE)
tweaked <- tweak(es.rf, newdata= scaled.X, label.from = "spam", label.to = "nonspam",
                 .dopar = TRUE)
plot(tweaked)
plot(tweaked, type="a")
plot(tweaked, type="d")
plot(tweaked, type="f")
}

}
