% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.eSatisfactory.R
\name{set.eSatisfactory}
\alias{set.eSatisfactory}
\title{get e-satisfactory instance of aim-leaf from all tree}
\usage{
set.eSatisfactory(rules, epsiron = 0.1)
}
\arguments{
\item{rules}{an objectrandomForest object.}

\item{epsiron}{a numeric. Amount of "tolerance". All standardized feature value changes from threashold with this tolerance.}
}
\value{
a list of trees (list).
}
\description{
get e-satisfactory instance of aim-leaf from all tree
}
\examples{
\dontrun{
X <- iris[, 1:(ncol(iris)-1)]
true.y <- iris[, ncol(iris)]

rf.iris <- randomForest(X, true.y, ntree=30)
rules.rf <- getRules(rf.iris, ktree = 20)
es.rf <- set.eSatisfactory(forest.rf, epsiron = 0.3)
}

}
