% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{tweak}
\alias{tweak}
\title{calcurate suggention for each inputs based on prediction}
\usage{
tweak(esrules, newdata, label.from, label.to, .dopar = TRUE)
}
\arguments{
\item{esrules}{an object returned by set.eSatisfactory().}

\item{newdata}{a data.frame or matrix. test data to be predicted and to be suggested. newdata must have the same structure of train data.}

\item{label.from}{a character. Predicted class label that user wants to change.}

\item{label.to}{a character. Class label that user wants to be changed from \code{label.from}.}

\item{.dopar}{logical. If \code{.dopar = TRUE}, suggestion for each instance will be calculated in parallel}
}
\value{
a list of trees (list).
}
\description{
calcurate suggention for each inputs based on prediction
}
\examples{
\dontrun{
X <- iris[, 1:(ncol(iris)-1)]
true.y <- iris[, ncol(iris)]

rf.iris <- randomForest(X, true.y, ntree=30)
es.rf <- set.eSatisfactory(forest.rf, ntree = 30, epsiron = 0.3, resample = TRUE)
}

}
