% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak.R
\name{tweak}
\alias{tweak}
\title{Calcurate suggention for each inputs based on prediction}
\usage{
tweak(esrules, newdata, label.from, label.to, .dopar = TRUE)
}
\arguments{
\item{esrules}{an object returned by set.eSatisfactory().}

\item{newdata}{a data.frame or matrix. test data to be predicted and to be suggested. newdata must have the same structure of train data.}

\item{label.from}{a character. Predicted class label that user wants to change.}

\item{label.to}{a character. Class label that user wants to be changed from \code{label.from}.}

\item{.dopar}{logical. If \code{.dopar = TRUE}, suggestion for each instance will be calculated in parallel}
}
\value{
a list of
\itemize{
 \item{"predict"}{ character vector of predicted label of each instance. See \code{\link{predict.randomForest}}. }
 \item{"original"}{ is the same as \code{newdata}. }
 \item{"suggest"}{ a data.frame of instances tweakd from newdata. }
}
}
\description{
Calcurate suggention for each inputs based on prediction
}
\examples{
\dontrun{
X <- iris[, 1:(ncol(iris)-1)]
scaled.X <- scale(X)
true.y <- iris[, ncol(iris)]

rf.iris <- randomForest(scaled.X, true.y, ntree=30)
es.rf <- set.eSatisfactory(forest.rf, ntree = 30, epsiron = 0.3, resample = TRUE)
tweaked <- tweak(es.rf, newdata= scaled.X, label.from = "spam", label.to = "nonspam",
                 .dopar = TRUE)
}

}
